Prevent cookie from being written in the wrong folder when using cron
The previous method worked but was lacking consistency. For instance if
the script was launched from a terminal, the cookie was created in the
right folder (e.g. '/home/user/hotspot-connect/.cookie'). However, when
the script was launched at boot with a cronjob, the cookie was created
in the user's home folder instead (e.g. '/home/user/.cookie').